requirements.txt
python-telegram-bot==22.5
langchain==0.3.27
langchain-xai==0.2.5
langchain-core==0.3.76
